Define YAMLSerializerField in its own file
