+++ src/aap_eda/api/serializers/event_stream.py
@@ -30,11 +30,8 @@ from aap_eda.api.exceptions import (
     MissingEventStreamRulebookSource,
 )
 from aap_eda.api.serializers.credential import CredentialSerializer
-from aap_eda.api.serializers.utils import (
-    YAMLSerializerField,
-    substitute_extra_vars,
-    swap_sources,
-)
+from aap_eda.api.serializers.fields.yaml import YAMLSerializerField
+from aap_eda.api.serializers.utils import substitute_extra_vars, swap_sources
 from aap_eda.core import models, validators
 
 logger = logging.getLogger(__name__)
